Move the pre-commit tests to a subdir
