+++ tests/functional/pre_commit/test_pre_commit_1_sync_with_config.py
@@ -5,7 +5,7 @@ from pre_commit.main import main as pre_commit
 
 from jupytext import TextFileContentsManager, read
 
-from .utils import (
+from ...utils import (
     skip_pre_commit_tests_on_windows,
     skip_pre_commit_tests_when_jupytext_folder_is_not_a_git_repo,
 )
