+++ tests/functional/pre_commit/test_pre_commit_2_sync_nbstripout.py
@@ -3,8 +3,9 @@ from git.exc import HookExecutionError
 from pre_commit.main import main as pre_commit
 
 from jupytext import read, write
+from jupytext.cli import jupytext
 
-from .utils import (
+from ...utils import (
     skip_pre_commit_tests_on_windows,
     skip_pre_commit_tests_when_jupytext_folder_is_not_a_git_repo,
 )
@@ -12,7 +13,7 @@ from .utils import (
 
 @skip_pre_commit_tests_on_windows
 @skip_pre_commit_tests_when_jupytext_folder_is_not_a_git_repo
-def test_pre_commit_hook_sync_black_nbstripout(
+def test_pre_commit_hook_sync_nbstripout(
     tmpdir,
     cwd_tmpdir,
     tmp_repo,
@@ -20,43 +21,35 @@ def test_pre_commit_hook_sync_black_nbstripout(
     jupytext_repo_rev,
     notebook_with_outputs,
 ):
-    """Here we sync the ipynb notebook with a py:percent file and also apply black and nbstripout."""
+    """Here we sync the ipynb notebook with a Markdown file and also apply nbstripout."""
     pre_commit_config_yaml = f"""
 repos:
 - repo: {jupytext_repo_root}
   rev: {jupytext_repo_rev}
   hooks:
   - id: jupytext
-    args: [--sync, --pipe, black]
-    additional_dependencies:
-    - black==22.3.0  # Matches hook
-
-- repo: https://github.com/psf/black
-  rev: 22.3.0
-  hooks:
-  - id: black
+    args: [--sync]
 
 - repo: https://github.com/kynan/nbstripout
   rev: 0.5.0
   hooks:
   - id: nbstripout
 """
+
     tmpdir.join(".pre-commit-config.yaml").write(pre_commit_config_yaml)
 
     tmp_repo.git.add(".pre-commit-config.yaml")
     pre_commit(["install", "--install-hooks", "-f"])
 
-    tmpdir.join("jupytext.toml").write('formats = "ipynb,py:percent"')
-    tmp_repo.git.add("jupytext.toml")
-    tmp_repo.index.commit("pair notebooks")
-
     # write a test notebook
     write(notebook_with_outputs, "test.ipynb")
 
+    # We pair the notebook to a md file
+    jupytext(["--set-formats", "ipynb,md", "test.ipynb"])
+
     # try to commit it, should fail because
-    # 1. the py version hasn't been added
-    # 2. the first cell is '1+1' which is not black compliant
-    # 3. the notebook has outputs
+    # 1. the md version hasn't been added
+    # 2. the notebook has outputs
     tmp_repo.git.add("test.ipynb")
     with pytest.raises(
         HookExecutionError,
@@ -66,16 +59,13 @@ repos:
 
     # Add the two files
     tmp_repo.git.add("test.ipynb")
-    tmp_repo.git.add("test.py")
+    tmp_repo.git.add("test.md")
 
     # now the commit will succeed
     tmp_repo.index.commit("passing")
     assert "test.ipynb" in tmp_repo.tree()
-    assert "test.py" in tmp_repo.tree()
+    assert "test.md" in tmp_repo.tree()
 
-    # the first cell was reformatted
+    # the ipynb file has no outputs on disk
     nb = read("test.ipynb")
-    assert nb.cells[0].source == "1 + 1"
-
-    # the ipynb file has no outputs
     assert not nb.cells[0].outputs
