+++ tests/functional/pre_commit/test_pre_commit_3_sync_black_nbstripout.py
@@ -1,21 +1,18 @@
 import pytest
 from git.exc import HookExecutionError
-from nbformat.v4.nbbase import new_code_cell
 from pre_commit.main import main as pre_commit
 
 from jupytext import read, write
 
-from .utils import (
-    requires_user_kernel_python3,
+from ...utils import (
     skip_pre_commit_tests_on_windows,
     skip_pre_commit_tests_when_jupytext_folder_is_not_a_git_repo,
 )
 
 
-@requires_user_kernel_python3
 @skip_pre_commit_tests_on_windows
 @skip_pre_commit_tests_when_jupytext_folder_is_not_a_git_repo
-def test_pre_commit_hook_sync_execute(
+def test_pre_commit_hook_sync_black_nbstripout(
     tmpdir,
     cwd_tmpdir,
     tmp_repo,
@@ -23,34 +20,43 @@ def test_pre_commit_hook_sync_execute(
     jupytext_repo_rev,
     notebook_with_outputs,
 ):
-    """Here we sync the ipynb notebook with a py:percent file and execute it (this is probably not a very
-    recommendable hook!)"""
+    """Here we sync the ipynb notebook with a py:percent file and also apply black and nbstripout."""
     pre_commit_config_yaml = f"""
 repos:
 - repo: {jupytext_repo_root}
   rev: {jupytext_repo_rev}
   hooks:
   - id: jupytext
-    args: [--sync, --execute, --show-changes]
+    args: [--sync, --pipe, black]
     additional_dependencies:
-    - nbconvert
+    - black==22.3.0  # Matches hook
+
+- repo: https://github.com/psf/black
+  rev: 22.3.0
+  hooks:
+  - id: black
+
+- repo: https://github.com/kynan/nbstripout
+  rev: 0.5.0
+  hooks:
+  - id: nbstripout
 """
     tmpdir.join(".pre-commit-config.yaml").write(pre_commit_config_yaml)
 
     tmp_repo.git.add(".pre-commit-config.yaml")
     pre_commit(["install", "--install-hooks", "-f"])
 
-    # simple notebook with kernel 'user_kernel_python3'
-    nb = notebook_with_outputs
-    nb.cells = [new_code_cell("3+4")]
+    tmpdir.join("jupytext.toml").write('formats = "ipynb,py:percent"')
+    tmp_repo.git.add("jupytext.toml")
+    tmp_repo.index.commit("pair notebooks")
 
-    # pair it to a py file and write it to disk
-    nb.metadata["jupytext"] = {"formats": "ipynb,py:percent"}
-    write(nb, "test.ipynb")
+    # write a test notebook
+    write(notebook_with_outputs, "test.ipynb")
 
     # try to commit it, should fail because
     # 1. the py version hasn't been added
-    # 2. the outputs are missing
+    # 2. the first cell is '1+1' which is not black compliant
+    # 3. the notebook has outputs
     tmp_repo.git.add("test.ipynb")
     with pytest.raises(
         HookExecutionError,
@@ -67,6 +73,9 @@ repos:
     assert "test.ipynb" in tmp_repo.tree()
     assert "test.py" in tmp_repo.tree()
 
-    # the first cell has the expected output
+    # the first cell was reformatted
     nb = read("test.ipynb")
-    assert nb.cells[0].outputs[0]["data"]["text/plain"] == "7"
+    assert nb.cells[0].source == "1 + 1"
+
+    # the ipynb file has no outputs
+    assert not nb.cells[0].outputs
