+++ tests/functional/pre_commit/test_pre_commit_4_sync_execute.py
@@ -1,19 +1,21 @@
 import pytest
 from git.exc import HookExecutionError
+from nbformat.v4.nbbase import new_code_cell
 from pre_commit.main import main as pre_commit
 
 from jupytext import read, write
-from jupytext.cli import jupytext
 
-from .utils import (
+from ...utils import (
+    requires_user_kernel_python3,
     skip_pre_commit_tests_on_windows,
     skip_pre_commit_tests_when_jupytext_folder_is_not_a_git_repo,
 )
 
 
+@requires_user_kernel_python3
 @skip_pre_commit_tests_on_windows
 @skip_pre_commit_tests_when_jupytext_folder_is_not_a_git_repo
-def test_pre_commit_hook_sync_nbstripout(
+def test_pre_commit_hook_sync_execute(
     tmpdir,
     cwd_tmpdir,
     tmp_repo,
@@ -21,35 +23,34 @@ def test_pre_commit_hook_sync_nbstripout(
     jupytext_repo_rev,
     notebook_with_outputs,
 ):
-    """Here we sync the ipynb notebook with a Markdown file and also apply nbstripout."""
+    """Here we sync the ipynb notebook with a py:percent file and execute it (this is probably not a very
+    recommendable hook!)"""
     pre_commit_config_yaml = f"""
 repos:
 - repo: {jupytext_repo_root}
   rev: {jupytext_repo_rev}
   hooks:
   - id: jupytext
-    args: [--sync]
-
-- repo: https://github.com/kynan/nbstripout
-  rev: 0.5.0
-  hooks:
-  - id: nbstripout
+    args: [--sync, --execute, --show-changes]
+    additional_dependencies:
+    - nbconvert
 """
-
     tmpdir.join(".pre-commit-config.yaml").write(pre_commit_config_yaml)
 
     tmp_repo.git.add(".pre-commit-config.yaml")
     pre_commit(["install", "--install-hooks", "-f"])
 
-    # write a test notebook
-    write(notebook_with_outputs, "test.ipynb")
+    # simple notebook with kernel 'user_kernel_python3'
+    nb = notebook_with_outputs
+    nb.cells = [new_code_cell("3+4")]
 
-    # We pair the notebook to a md file
-    jupytext(["--set-formats", "ipynb,md", "test.ipynb"])
+    # pair it to a py file and write it to disk
+    nb.metadata["jupytext"] = {"formats": "ipynb,py:percent"}
+    write(nb, "test.ipynb")
 
     # try to commit it, should fail because
-    # 1. the md version hasn't been added
-    # 2. the notebook has outputs
+    # 1. the py version hasn't been added
+    # 2. the outputs are missing
     tmp_repo.git.add("test.ipynb")
     with pytest.raises(
         HookExecutionError,
@@ -59,13 +60,13 @@ repos:
 
     # Add the two files
     tmp_repo.git.add("test.ipynb")
-    tmp_repo.git.add("test.md")
+    tmp_repo.git.add("test.py")
 
     # now the commit will succeed
     tmp_repo.index.commit("passing")
     assert "test.ipynb" in tmp_repo.tree()
-    assert "test.md" in tmp_repo.tree()
+    assert "test.py" in tmp_repo.tree()
 
-    # the ipynb file has no outputs on disk
+    # the first cell has the expected output
     nb = read("test.ipynb")
-    assert not nb.cells[0].outputs
+    assert nb.cells[0].outputs[0]["data"]["text/plain"] == "7"
