+++ tests/functional/pre_commit/test_pre_commit_mode.py
